move InList related simplify to one place
