+++ datafusion/optimizer/src/simplify_expressions/inlist_simplifier.rs
@@ -17,12 +17,169 @@
 
 //! This module implements a rule that simplifies the values for `InList`s
 
+use std::borrow::Cow;
 use std::collections::HashSet;
 
 use datafusion_common::tree_node::TreeNodeRewriter;
-use datafusion_common::Result;
-use datafusion_expr::expr::InList;
+use datafusion_common::{Result, ScalarValue};
+use datafusion_expr::expr::{InList, InSubquery};
 use datafusion_expr::{lit, BinaryExpr, Expr, Operator};
+
+use super::utils::{is_null, lit_bool_null};
+use super::THRESHOLD_INLINE_INLIST;
+
+pub(super) struct ShortenInListSimplifier {}
+
+impl ShortenInListSimplifier {
+    pub(super) fn new() -> Self {
+        Self {}
+    }
+}
+
+impl TreeNodeRewriter for ShortenInListSimplifier {
+    type N = Expr;
+
+    fn mutate(&mut self, expr: Expr) -> Result<Expr> {
+        if let Expr::InList(InList {
+            expr,
+            list,
+            negated,
+        }) = expr.clone()
+        {
+            if !list.is_empty()
+                && (
+                    // For lists with only 1 value we allow more complex expressions to be simplified
+                    // e.g SUBSTR(c1, 2, 3) IN ('1') -> SUBSTR(c1, 2, 3) = '1'
+                    // for more than one we avoid repeating this potentially expensive
+                    // expressions
+                    list.len() == 1
+                        || list.len() <= THRESHOLD_INLINE_INLIST
+                            && expr.try_into_col().is_ok()
+                )
+            {
+                let first_val = list[0].clone();
+                if negated {
+                    return Ok(list.into_iter().skip(1).fold(
+                        (*expr.clone()).not_eq(first_val),
+                        |acc, y| {
+                            // Note that `A and B and C and D` is a left-deep tree structure
+                            // as such we want to maintain this structure as much as possible
+                            // to avoid reordering the expression during each optimization
+                            // pass.
+                            //
+                            // Left-deep tree structure for `A and B and C and D`:
+                            // ```
+                            //        &
+                            //       / \
+                            //      &   D
+                            //     / \
+                            //    &   C
+                            //   / \
+                            //  A   B
+                            // ```
+                            //
+                            // The code below maintain the left-deep tree structure.
+                            acc.and((*expr.clone()).not_eq(y))
+                        },
+                    ));
+                } else {
+                    return Ok(list.into_iter().skip(1).fold(
+                        (*expr.clone()).eq(first_val),
+                        |acc, y| {
+                            // Same reasoning as above
+                            acc.or((*expr.clone()).eq(y))
+                        },
+                    ));
+                }
+            }
+        }
+
+        Ok(expr)
+    }
+}
+
+pub(super) struct InListSimplifier {}
+
+impl InListSimplifier {
+    pub(super) fn new() -> Self {
+        Self {}
+    }
+}
+
+impl TreeNodeRewriter for InListSimplifier {
+    type N = Expr;
+
+    fn mutate(&mut self, expr: Expr) -> Result<Expr> {
+        let mut new_expr = null_simplify(expr)?;
+        new_expr = or_inlist_simplify(new_expr)?;
+        new_expr = simplify(new_expr)?;
+
+        Ok(new_expr)
+    }
+}
+
+fn null_simplify(expr: Expr) -> Result<Expr> {
+    if let Expr::InList(InList {
+        expr,
+        mut list,
+        negated,
+    }) = expr.clone()
+    {
+        if list.is_empty() && *expr != Expr::Literal(ScalarValue::Null) {
+            return Ok(lit(negated));
+        } else if is_null(&expr) {
+            return Ok(lit_bool_null());
+        } else if list.len() == 1
+            && matches!(list.first(), Some(Expr::ScalarSubquery { .. }))
+        {
+            let Expr::ScalarSubquery(subquery) = list.remove(0) else {
+                unreachable!()
+            };
+            return Ok(Expr::InSubquery(InSubquery::new(expr, subquery, negated)));
+        }
+    }
+
+    Ok(expr)
+}
+
+/// Combine multiple OR expressions into a single IN list expression if possible
+///
+/// i.e. `a = 1 OR a = 2 OR a = 3` -> `a IN (1, 2, 3)`
+fn or_inlist_simplify(expr: Expr) -> Result<Expr> {
+    if let Expr::BinaryExpr(BinaryExpr { left, op, right }) = &expr {
+        if *op == Operator::Or {
+            let left = as_inlist(left);
+            let right = as_inlist(right);
+            if let (Some(lhs), Some(rhs)) = (left, right) {
+                if lhs.expr.try_into_col().is_ok()
+                    && rhs.expr.try_into_col().is_ok()
+                    && lhs.expr == rhs.expr
+                    && !lhs.negated
+                    && !rhs.negated
+                {
+                    let lhs = lhs.into_owned();
+                    let rhs = rhs.into_owned();
+                    let mut seen: HashSet<Expr> = HashSet::new();
+                    let list = lhs
+                        .list
+                        .into_iter()
+                        .chain(rhs.list)
+                        .filter(|e| seen.insert(e.to_owned()))
+                        .collect::<Vec<_>>();
+
+                    let merged_inlist = InList {
+                        expr: lhs.expr,
+                        list,
+                        negated: false,
+                    };
+                    return Ok(Expr::InList(merged_inlist));
+                }
+            }
+        }
+    }
+
+    Ok(expr)
+}
 
 /// Simplify expressions that is guaranteed to be true or false to a literal boolean expression
 ///
@@ -40,41 +197,52 @@ use datafusion_expr::{lit, BinaryExpr, Expr, Operator};
 ///     6. `a in (1,2,3,4) AND a not in (1,2,3,4,5) -> a in (), which is false`
 ///     7. `a not in (1,2,3,4) AND a in (1,2,3,4,5) -> a = 5`
 ///     8. `a in (1,2,3,4) AND a not in (5,6,7,8) -> a in (1,2,3,4)`
-pub(super) struct InListSimplifier {}
-
-impl InListSimplifier {
-    pub(super) fn new() -> Self {
-        Self {}
-    }
-}
-
-impl TreeNodeRewriter for InListSimplifier {
-    type N = Expr;
-
-    fn mutate(&mut self, expr: Expr) -> Result<Expr> {
-        if let Expr::BinaryExpr(BinaryExpr { left, op, right }) = &expr {
-            if let (Expr::InList(l1), Operator::And, Expr::InList(l2)) =
-                (left.as_ref(), op, right.as_ref())
-            {
-                if l1.expr == l2.expr && !l1.negated && !l2.negated {
-                    return inlist_intersection(l1, l2, false);
-                } else if l1.expr == l2.expr && l1.negated && l2.negated {
-                    return inlist_union(l1, l2, true);
-                } else if l1.expr == l2.expr && !l1.negated && l2.negated {
-                    return inlist_except(l1, l2);
-                } else if l1.expr == l2.expr && l1.negated && !l2.negated {
-                    return inlist_except(l2, l1);
-                }
-            } else if let (Expr::InList(l1), Operator::Or, Expr::InList(l2)) =
-                (left.as_ref(), op, right.as_ref())
-            {
-                if l1.expr == l2.expr && l1.negated && l2.negated {
-                    return inlist_intersection(l1, l2, true);
-                }
+fn simplify(expr: Expr) -> Result<Expr> {
+    if let Expr::BinaryExpr(BinaryExpr { left, op, right }) = &expr {
+        if let (Expr::InList(l1), Operator::And, Expr::InList(l2)) =
+            (left.as_ref(), op, right.as_ref())
+        {
+            if l1.expr == l2.expr && !l1.negated && !l2.negated {
+                return inlist_intersection(l1, l2, false);
+            } else if l1.expr == l2.expr && l1.negated && l2.negated {
+                return inlist_union(l1, l2, true);
+            } else if l1.expr == l2.expr && !l1.negated && l2.negated {
+                return inlist_except(l1, l2);
+            } else if l1.expr == l2.expr && l1.negated && !l2.negated {
+                return inlist_except(l2, l1);
+            }
+        } else if let (Expr::InList(l1), Operator::Or, Expr::InList(l2)) =
+            (left.as_ref(), op, right.as_ref())
+        {
+            if l1.expr == l2.expr && l1.negated && l2.negated {
+                return inlist_intersection(l1, l2, true);
             }
         }
+    }
 
-        Ok(expr)
+    Ok(expr)
+}
+
+/// Try to convert an expression to an in-list expression
+fn as_inlist(expr: &Expr) -> Option<Cow<InList>> {
+    match expr {
+        Expr::InList(inlist) => Some(Cow::Borrowed(inlist)),
+        Expr::BinaryExpr(BinaryExpr { left, op, right }) if *op == Operator::Eq => {
+            match (left.as_ref(), right.as_ref()) {
+                (Expr::Column(_), Expr::Literal(_)) => Some(Cow::Owned(InList {
+                    expr: left.clone(),
+                    list: vec![*right.clone()],
+                    negated: false,
+                })),
+                (Expr::Literal(_), Expr::Column(_)) => Some(Cow::Owned(InList {
+                    expr: right.clone(),
+                    list: vec![*left.clone()],
+                    negated: false,
+                })),
+                _ => None,
+            }
+        }
+        _ => None,
     }
 }
