+++ datafusion/optimizer/src/simplify_expressions/mod.rs
@@ -19,7 +19,6 @@ pub mod context;
 pub mod expr_simplifier;
 mod guarantees;
 mod inlist_simplifier;
-mod or_in_list_simplifier;
 mod regex;
 pub mod simplify_exprs;
 mod utils;
